Move all RefRefMapWrappers into one class
Now:
RefRefMapWrapper.FromMap
RefRefMapWrapper.FromObjectRefMap
RefRefMapWrapper.FromRefObjectMap
